Move systems to their own module.
Since systems tend to have a broad *supervisory* view of the game state, it
makes some sense to have them all in a module space of their own, as opposed
to employing the fractal pattern, pairing systems up with component
definitions.
